fix fifo
