+++ tunikov.dmitrii/lab2/conn/conn_mq.cpp
@@ -10,16 +10,22 @@
 #include <cstring>
 #include <mqueue.h>
 
-int Conn::desc;
-const char* Conn::channel_name;
-
-Conn::Conn()
+Conn::Conn() : is_open(false)
 {
+    desc = new (std::nothrow) int (-1);
     channel_name = "/mq_queue";
+}
+
+Conn::~Conn()
+{
+    delete desc;
 }
 
 bool Conn::Open(size_t id, bool create)
 {
+    if (desc == nullptr)
+        return false;
+
     bool res = false;
     is_host = create;
     int mqflg = O_RDWR;
@@ -30,31 +36,38 @@ bool Conn::Open(size_t id, bool create)
         std::cout << "Creating connection with id = " << id << std::endl;
         mqflg |= O_CREAT;
         struct mq_attr attr = ((struct mq_attr){0, 1, sizeof(Message), 0, {0}});
-        desc = mq_open(channel_name, mqflg, mqperm, &attr);
+        *desc = mq_open(channel_name, mqflg, mqperm, &attr);
     }
     else
     {
         std::cout << "Getting connection with id = " << id << std::endl;
-        desc = mq_open(channel_name, mqflg);
+        *desc = mq_open(channel_name, mqflg);
     }
 
-    if (desc == -1)
+    if (*desc == -1)
     {
         std::cout << "ERROR: mq_open failed, errno = " << strerror(errno) << std::endl;
     }
     else
     {
-        std::cout << "mq_open returned id = " << desc << std::endl;
+        std::cout << "mq_open returned id = " << *desc << std::endl;
         res = true;
     }
+    is_open = *desc != -1;
     return res;
 }
 
 bool Conn::Read(void* buf, size_t count)
 {
+    if (!is_open)
+    {
+        std::cout << "ERROR: can't read bacause mq not opened" << std::endl;
+        return false;
+    }
+
     Message mq_buf;
     bool success = true;
-    if (mq_receive(desc, (char *)&mq_buf, sizeof(Message), nullptr) == -1)
+    if (mq_receive(*desc, (char *)&mq_buf, sizeof(Message), nullptr) == -1)
     {
         std::cout << "ERROR: mq_recieve failed, errno = " << strerror(errno) << std::endl;
         success = false;
@@ -68,10 +81,16 @@ bool Conn::Read(void* buf, size_t count)
 
 bool Conn::Write(void* buf, size_t count)
 {
+    if (!is_open)
+    {
+        std::cout << "ERROR: can't read bacause mq not opened" << std::endl;
+        return false;
+    }
+
     bool res = false;
     if (count <= sizeof(Message))
     {
-        if (mq_send(desc, (char*)buf, count, 0) == -1)
+        if (mq_send(*desc, (char*)buf, count, 0) == -1)
         {
             std::cout << "ERROR: mq_send failed, errno = " << strerror(errno) << std::endl;
         }
@@ -85,8 +104,14 @@ bool Conn::Write(void* buf, size_t count)
 
 bool Conn::Close()
 {
+    if (!is_open)
+    {
+        std::cout << "ERROR: can't read bacause mq not opened" << std::endl;
+        return false;
+    }
+
     bool res = false;
-    if (mq_close(desc) == 0)
+    if (mq_close(*desc) == 0)
     {
         if (!is_host)
         {
